Clean/dedupe integration tests
- Create a base-level 'testing' directory with common SDK-specific utils (layer above shakedown)
- Move common bits into 'testing' directory and update per-project tests to use it
- Flatten directory structure in frameworks: just '/tests/' instead of '/integration/tests'
- Remove unexercised tests in kafka (appear to just be copied from hdfs and never run)
