+++ frameworks/hdfs/tests/requirements.txt
