+++ frameworks/hdfs/tests/strict.json
