+++ frameworks/helloworld/tests/test_sanity.py
@@ -5,19 +5,7 @@ import pytest
 import re
 import shakedown
 
-from tests.test_utils import (
-    PACKAGE_NAME,
-    check_health,
-    get_marathon_config,
-    get_deployment_plan,
-    get_task_count,
-    install,
-    marathon_api_url,
-    request,
-    run_dcos_cli_cmd,
-    uninstall,
-    spin
-)
+PACKAGE_NAME = 'hello-world'
 
 
 def setup_module(module):
@@ -225,66 +213,6 @@ def test_lock():
     assert zk_config_old == zk_config_new
 
 
-def get_task_ids(prefix):
-    tasks = shakedown.get_service_tasks(PACKAGE_NAME)
-    prefixed_tasks = [t for t in tasks if t['name'].startswith(prefix)]
-    task_ids = [t['id'] for t in prefixed_tasks]
-    return task_ids
-
-
-def tasks_updated(prefix, old_task_ids):
-    def fn():
-        try:
-            return get_task_ids(prefix)
-        except dcos.errors.DCOSHTTPException:
-            return []
-
-    def success_predicate(task_ids):
-        print('Old task ids: ' + str(old_task_ids))
-        print('New task ids: ' + str(task_ids))
-        success = True
-
-        for id in task_ids:
-            print('Checking ' + id)
-            if id in old_task_ids:
-                success = False
-
-        if not len(task_ids) >= len(old_task_ids):
-            success = False
-
-        print('Waiting for update to ' + prefix)
-        return (
-            success,
-            'Task type:' + prefix + ' not updated'
-        )
-
-    return spin(fn, success_predicate)
-
-
-def tasks_not_updated(prefix, old_task_ids):
-    def fn():
-        try:
-            return get_task_ids(prefix)
-        except dcos.errors.DCOSHTTPException:
-            return []
-
-    def success_predicate(task_ids):
-        print('Old task ids: ' + str(old_task_ids))
-        print('New task ids: ' + str(task_ids))
-        success = True
-
-        for id in old_task_ids:
-            print('Checking ' + id)
-            if id not in task_ids:
-                success = False
-
-        if not len(task_ids) >= len(old_task_ids):
-            success = False
-
-        print('Determining no update occurred for ' + prefix)
-        return (
-            success,
-            'Task type:' + prefix + ' not updated'
-        )
-
-    return spin(fn, success_predicate)
+def get_task_count():
+    config = sdk_marathon.get_marathon_config()
+    return int(config['env']['HELLO_COUNT']) + int(config['env']['WORLD_COUNT'])
