+++ frameworks/helloworld/tests/test_web_url.py
@@ -4,21 +4,7 @@ import pytest
 import re
 import shakedown
 
-from tests.test_utils import (
-    PACKAGE_NAME,
-    check_health,
-    get_marathon_config,
-    get_deployment_plan,
-    get_sidecar_plan,
-    get_task_count,
-    install,
-    marathon_api_url,
-    request,
-    run_dcos_cli_cmd,
-    uninstall,
-    spin,
-    start_sidecar_plan
-)
+PACKAGE_NAME = 'hello-world'
 
 
 def setup_module(module):
