--- frameworks/kafka/integration/tests/strict.json
@@ -1,6 +1,0 @@
-{
-    "service": {
-        "principal": "service-acct",
-        "secret_name": "secret"
-    }
-}
