+++ test.sh
@@ -25,15 +25,15 @@ function run_framework_tests {
         export STUB_UNIVERSE_URL=$(cat $UNIVERSE_URL_PATH)
         rm -f $UNIVERSE_URL_PATH
         echo "Built/uploaded stub universe: $STUB_UNIVERSE_URL"
-    else 
+    else
         echo "Using provided STUB_UNIVERSE_URL: $STUB_UNIVERSE_URL"
     fi
 
     # Run shakedown tests in framework scheduler directory:
     ${REPO_ROOT_DIR}/tools/run_tests.py \
                     shakedown \
-                    ${FRAMEWORK_DIR}/integration/tests/ \
-                    ${FRAMEWORK_DIR}/integration/requirements.txt
+                    ${FRAMEWORK_DIR}/tests/ \
+                    ${FRAMEWORK_DIR}/tests/requirements.txt
 }
 
 REPO_ROOT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
