+++ testing/README.md
@@ -1,0 +1,3 @@
+# Testing utils
+
+Common python code used in the [shakedown](https://github.com/mesosphere/shakedown)-based integration tests for services in this repository.
